requirements.txt
# Core dependencies
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0

# Audio processing
SpeechRecognition>=3.10.0
pygame>=2.5.0
gTTS>=2.4.0
PyAudio>=0.2.11

# AI and async
aiohttp>=3.8.0
asyncio

# Enhanced gesture detection (optional)
mediapipe>=0.10.0

# Windows TTS (optional)
pywin32>=306; sys_platform == "win32"

# Image processing
Pillow>=10.0.0

# Logging and utilities
python-dateutil>=2.8.0

# Development and testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0 